Don't show New choser at startup.  Rename Save Js and Html to Save Html (and comment out call to save js).  Rename template folders so users don't access them.
Added a space (&nbsp;) after the numbering in Fill-Ins

Renamed Template folders so easier for user to understand never to open them.


Feedback in Multiple Choice quizzes behaves opposite of Questions and Answers (since it's not shown in html but in alerts and textboxes) so we need to convert &quot; to \\\" when creating js.  And, we need to convert \\\" to &quot when filling into the datagrid.